refactor: refactoring FakeListAPI
